B-因子构建类/机器学习/基于机器学习的AlphaGPT因子挖掘/pyproject.toml
[project]
name = "alphagpt-ashare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pyqlib>=0.9.7",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]
